refactor: controller code refactor and js edit (#51)
