+++ src/main/java/like/lion/way/feed/controller/CommentRestController.java
@@ -9,9 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -19,20 +17,18 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class CommentRestController {
     private final PostCommentService postCommentService;
-
+    //댓글 수정
     @PatchMapping("/posts/comments/{commentId}")
     public ResponseEntity<String> updateComment(@PathVariable Long commentId, @RequestBody Map<String, String> payload) {
         try {
             String updatedContent = payload.get("updatedContent");
-            log.info("commentId: {}", commentId);
-            log.info("content: {}", updatedContent);
             postCommentService.updateComment(commentId, updatedContent);
             return ResponseEntity.ok("comment updated successfully.");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to update comment.");
         }
     }
-
+    //댓글 삭제
     @DeleteMapping("/posts/comments/{commentId}")
     public ResponseEntity<String> deleteComment(@PathVariable Long commentId) {
         try {
