+++ src/main/java/like/lion/way/feed/controller/GetUserIp.java
@@ -1,0 +1,25 @@
+package like.lion.way.feed.controller;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public class GetUserIp {
+    public static String getRemoteIP(HttpServletRequest request){
+        String ip = request.getHeader("X-FORWARDED-FOR");
+
+        //proxy 환경일 경우
+        if (ip == null || ip.length() == 0) {
+            ip = request.getHeader("Proxy-Client-IP");
+        }
+
+        //웹로직 서버일 경우
+        if (ip == null || ip.length() == 0) {
+            ip = request.getHeader("WL-Proxy-Client-IP");
+        }
+
+        if (ip == null || ip.length() == 0) {
+            ip = request.getRemoteAddr() ;
+        }
+
+        return ip;
+    }
+}
