+++ src/main/java/like/lion/way/feed/controller/Post2Controller.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.time.LocalDateTime;
 import like.lion.way.feed.domain.Post;
 import like.lion.way.feed.domain.dto.PostDto;
+import like.lion.way.feed.service.PostBoxService;
 import like.lion.way.feed.service.PostService;
 import like.lion.way.jwt.util.JwtUtil;
 import like.lion.way.user.domain.User;
@@ -14,6 +15,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -26,9 +29,7 @@ public class Post2Controller {
     private final PostService postService;
     private final JwtUtil jwtUtil;
     private final UserService userService;
-
-    @Value("${image.upload.dir}")
-    private String uploadDir;
+    private final PostBoxService postBoxService;
 
     // 로그인한 사용자 정보 조회
     private User getLoginUser(HttpServletRequest request) {
@@ -39,7 +40,7 @@ public class Post2Controller {
         Long loginId = jwtUtil.getUserIdFromToken(token);
         return userService.findByUserId(loginId);
     }
-
+    //고정 핀 설정
     @PostMapping("/posts/pin/{postId}")
     public String pinPost(@PathVariable("postId") Long postId) {
         postService.pinPost(postId);
@@ -49,29 +50,28 @@ public class Post2Controller {
     // 게시판 생성
     @PostMapping("/posts/create")
     public String savePost(PostDto postDto, @RequestPart(value = "image") MultipartFile file, HttpServletRequest request){
-        Post post = new Post();
-        post.setPostTitle(postDto.getTitle()); // 제목
-        post.setPostContent(postDto.getContent()); // 내용
-
-        // 이미지 파일 저장
-        if (!file.isEmpty()) {
-            try {
-                String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
-                String filePath = uploadDir + File.separator + fileName;
-                File dest = new File(filePath);
-                file.transferTo(dest);
-                post.setPostImageUrl(fileName); // 웹에서 접근할 경로
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
         User user = getLoginUser(request);
-        log.info("username::::" + user.getUsername());
-        post.setUser(user); // 작성자 설정
-        post.setPostCreatedAt(LocalDateTime.now()); // 작성일
-        post.setPostLike(0); // 좋아요 수
-        postService.savePost(post);
+        postService.savePost(postDto, file, user);
         return "redirect:/posts";
+    }
+    // 게시판 생성 페이지로 넘어감
+    @GetMapping("/posts/create")
+    public String createPost() {
+        return "/pages/feed/feedCreate";
+    }
+
+
+    // 게시판 상세 (게시판 == 피드)
+    @GetMapping("/posts/detail/{postId}")
+    public String showDetailPost(@PathVariable("postId") Long postId, Model model, HttpServletRequest request){
+        log.info("postId:::: " + postId);
+        Post post = postService.getPostById(postId);
+        model.addAttribute("post", post);
+
+        model.addAttribute("postBox", postBoxService.getPostBoxByPostId(post).size());
+
+        User loginUser = getLoginUser(request);
+        model.addAttribute("loginUser", loginUser);
+        return "/pages/feed/detailFeed";
     }
 }
