+++ src/main/java/like/lion/way/feed/controller/Question2Controller.java
@@ -11,11 +11,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 @RequiredArgsConstructor
 public class Question2Controller {
     private final QuestionService questionService;
+    //질문 고정
     @PostMapping("/questions/pin/{questionId}")
     public String pinQuestion(@PathVariable("questionId") Long questionId) {
         questionService.pinQuestion(questionId);
         return "redirect:/questions/create";
     }
+    //질문 삭제
     @PostMapping("/questions/delete")
     public String deleteQuestion(@RequestParam("questionId") Long questionId) {
         questionService.deleteQuestion(questionId);
