+++ src/main/java/like/lion/way/feed/service/PostService.java
@@ -3,13 +3,13 @@ package like.lion.way.feed.service;
 import java.util.List;
 import like.lion.way.feed.domain.Post;
 import like.lion.way.feed.domain.PostBox;
+import like.lion.way.feed.domain.dto.PostDto;
 import like.lion.way.user.domain.User;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public interface PostService {
-    Post savePost(Post post);
-
     List<Post> getAllPosts();
 
     Post getPostById(Long id);
@@ -21,4 +21,6 @@ public interface PostService {
     List<Post> getPostByUser(User user);
 
     Post pinPost(Long postId);
+
+    Post savePost(PostDto postDto, MultipartFile file, User user);
 }
