+++ src/main/java/like/lion/way/feed/service/QuestionService.java
@@ -1,13 +1,14 @@
 package like.lion.way.feed.service;
 
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import like.lion.way.feed.domain.Question;
 import like.lion.way.user.domain.User;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public interface QuestionService {
-    Question saveQuestion(Question newQuestion);
 
     List<Question> getAllQuestions();
 
@@ -24,4 +25,8 @@ public interface QuestionService {
     Question rejectedQuestion(Question question);
 
     void deleteQuestion(Long questionId);
+
+    Question saveQuestion(User user, Long userId, String question, boolean isAnonymous, MultipartFile image, HttpServletRequest request);
+
+    Question saveQuestion(Question question, String answer);
 }
